requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
yfinance==0.2.22
gunicorn==21.2.0
python-telegram-bot==20.6
python-dotenv==1.0.0
